quantum_state_anchoring_library/src/state_anchoring.py

import numpy as np
from datetime import datetime, timedelta
from qiskit import QuantumCircuit, execute, Aer
from qiskit.quantum_info import Statevector

class QuantumStateAnchor:
    def __init__(self, quantum_circuit: QuantumCircuit, anchor_time: datetime = None):
        self.quantum_circuit = quantum_circuit
        self.anchor_time = anchor_time if anchor_time else datetime.utcnow()
        self.statevector = None

    def initialize_state(self):
        """Initialize the quantum state for anchoring"""
        backend = Aer.get_backend('statevector_simulator')
        job = execute(self.quantum_circuit, backend)
        result = job.result()
        self.statevector = Statevector(result.get_statevector(self.quantum_circuit))

    def anchor_state(self, anchor_time: datetime = None):
        """Anchor the quantum state at a given time"""
        if anchor_time:
            self.anchor_time = anchor_time
        else:
            self.anchor_time = datetime.utcnow()
        self.initialize_state()
        return {
            'statevector': self.statevector,
            'anchor_time': self.anchor_time.isoformat()
        }

    def compare_states(self, new_statevector: Statevector, threshold: float = 1e-5):
        """Compare current anchored state with a new statevector"""
        fidelity = self.statevector.fidelity(new_statevector)
        return fidelity > (1 - threshold)

    def time_travel(self, forward_seconds: int):
        """Simulate the passage of time and apply corrections if necessary"""
        future_time = self.anchor_time + timedelta(seconds=forward_seconds)
        self.apply_time_dilation(future_time)
        return future_time

    def apply_time_dilation(self, new_time: datetime):
        """Apply time dilation to the quantum state based on relativistic effects"""
        time_diff = (new_time - self.anchor_time).total_seconds()
        if time_diff > 0:
            self.quantum_circuit.rx(time_diff * np.pi / 4, 0)  # Adjust qubit rotation based on time shift
        self.anchor_time = new_time

    def save_state(self, file_path: str):
        """Save the current quantum state to a file"""
        with open(file_path, 'w') as f:
            f.write(f'Anchor Time: {self.anchor_time.isoformat()}\n')
            f.write(f'Statevector: {self.statevector}\n')

    def load_state(self, file_path: str):
        """Load a quantum state from a file"""
        with open(file_path, 'r') as f:
            lines = f.readlines()
            anchor_time_str = lines[0].split(":")[1].strip()
            statevector_str = lines[1].split(":")[1].strip()
            self.anchor_time = datetime.fromisoformat(anchor_time_str)
            self.statevector = Statevector(eval(statevector_str))

    def verify_anchored_state(self):
        """Verify the integrity of the anchored quantum state"""
        assert self.statevector is not None, "Statevector not initialized"
        assert self.anchor_time is not None, "Anchor time not set"
        return True
